backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
spacy
openai
python-dotenv
httpx